Activation function list

This data list the example of the activation function in RNN-based model

data("Activation_info")

Format

A data frame with 277 observations on the following 6 variables.

X

a numeric vector

Number

a numeric vector

Activation_name

a factor with levels Act_ART1_NC Act_ART2_Identity Act_ART2_NormIP Act_ART2_NormP Act_ART2_NormV Act_ART2_NormW Act_ART2_Rec Act_ART2_Rst Act_ARTMAP_DRho Act_ARTMAP_NCa Act_ARTMAP_NCb Act_at_least_1 Act_at_least_2 Act_at_most_0 Act_BAM Act_BSB Act_CC_Thresh Act_Component Act_Elliott Act_Euclid Act_exactly_1 Act_Exponential Act_HystStep Act_Identity Act_IdentityPlusBias Act_less_than_0 Act_Logistic Act_LogisticTbl Act_LogSym Act_MinOutPlusWeight Act_Perceptron Act_Product Act_RBF_Gaussian Act_RBF_MultiQuadratic Act_RBF_ThinPlateSpline Act_RM Act_Signum Act_Signum0 Act_Sinus Act_Softmax Act_StepFunc Act_TACOMA Act_TanH Act_TanHPlusBias Act_TanH_Xdiv2 Act_TD_Elliott Act_TD_Logistic ART1 ART1_Stable ART1_Synchronous ART1_Weights ART2 ART2_Stable ART2_Synchronous ART2_Weights ARTMAP ARTMAP_Stable ARTMAP_Synchronous ARTMAP_Weights Auto_Synchronous BackPercolation BackpropBatch BackpropChunk BackpropClassJogChunk BackpropJogChunk BackpropMomentum BackpropWeightDecay BAM_Order BBPTT Binary BPTT BPTT_Order CC CC_Order CC_Weights Clip ClippHebb Counterpropagation CounterPropagation CPN_Rand_Pat CPN_Weights_v3.2 CPN_Weights_v3.3 DLVQ_Weights Dynamic_LVQ ENZO_noinit ENZO_prop Hebb Hebb_Fixed_Act Hebbian Hopfield_Fixed_Act Hopfield_Synchronous Inverse JE_BP JE_BP_Momentum JE_Order JE_Quickprop JE_Rprop JE_Special JE_Weights Kohonen Kohonen_Const Kohonen_Order Kohonen_Rand_Pat Kohonen_Weights_v3.2 LinearScale Logistic_notInhibit MagPruning Monte-Carlo Noncontributing_Units None Norm OptimalBrainDamage OptimalBrainSurgeon Out_ART2_Noise_ContDiff Out_ART2_Noise_PLin Out_Clip_01 Out_Clip_11 Out_Identity Out_Threshold05 PruningFeedForward PseudoInv QPTT Quickprop RadialBasisLearning Randomize_Weights Random_Order Random_Permutation Random_Weights_Perc RBF-DDA RBF_Weights RBF_Weights_Kohonen RBF_Weights_Redo RM_delta RM_Random_Weights Rprop RpropMAP SCG Serial_Order Sim_Ann_SS Sim_Ann_WTA Sim_Ann_WWTA Site_at_least_1 Site_at_least_2 Site_at_most_0 Site_Max Site_Min Site_Pi Site_Produkt Site_Reciprocal Site_WeightedSum Skeletonization Std_Backpropagation Synchonous_Order TACOMA Threshold TimeDelayBackprop TimeDelay_Order Topological_Order

type

a numeric vector

Number_of_inParams

a numeric vector

Number_of_outParams

a numeric vector

Details

Example of available activation function in RNNS based models

Source

References

Bergmeir, C. N., & Benítez Sánchez, J. M. (2012). Neural networks in R using the Stuttgart neural network simulator: RSNNS. American Statistical Association.