Estimating Garson weights from FCNN4R models

This function estimates the weights of loci based on Garson et al.(2013).

GarsonWeights_FCNN4R(object, xnames = NULL, ynames = NULL)

Arguments

object
FCNN4R model

xnames
optional, independent variable names

ynames
optional, dependent variable names

Details

This function estimates the weights of neurons/loci based on Garson et al.(2013). Based on Gevrey et al. (2003), for each hidden neuron i, multiply the absolute value of the hidden-output layer connection weight by the absolute value of the hidden-input layer connection weight. Do this for each input variable j. The following products Pij are obtained.

Value

an array with the weights of each locus.

References

Gevrey, M., Dimopoulos, I., & Lek, S. (2003). Review and comparison of methods to study the contribution of variables in artificial neural network models. Ecological modelling, 160(3), 249-264.