Available models in DeepGenomeScan

The model list that shows the available model in DeepGenomeScan

data("model_info")

Format

A data frame with 35 observations on the following 6 variables.

X

a numeric vector

name

a factor with levels Convolutional Neural Network Deep learning (Multi-Layer Perceptron, MLP) glmnet (Lasso and Elastic-Net Regularized Generalized Models) Linear Regression Logic Regression Model Averaged Neural Network Monotone Multi-Layer Perceptron Neural Network Multi-Layer Perceptron (MLP) Multi-Layer Perceptron, MLP, multiple layers Multilayer Perceptron Network by Stochastic Gradient Descent Multilayer Perceptron Network with Dropout Multilayer Perceptron Network with Weight Decay Multi-Layer Perceptron Neural Network Multi-Layer Perceptron, with multiple layers Multi-Step Adaptive MCP-Net Neural Network Neural Networks with Feature Extraction Principal Component Analysis Quantile Regression Neural Network Ridge Regression Ridge Regression with Variable Selection Stacked AutoEncoder Deep Neural Network Stochastic Gradient Boosting The Bayesian lasso The lasso Tree Models from Genetic Algorithms

model

a factor with levels avNNet blasso CNNsgd dnn evtree foba gbm glmnet glmnet_h2o lasso lm logreg mlp mlpFCNN4Rsgd mlph2o mlpKerasDecay mlpKerasDropout mlpML mlpneuralnet mlpneuralnet1 mlpSGD mlpWeightDecay mlpWeightDecayML modelmlpkerasdropout modelRSNNSmlpdecay monmlp msaenet mxnet mxnetAdam neuralnet nnet pcaNNet pcr qrnn ridge

type

a factor with levels Classification, Regression Regression

libraries

a factor with levels deepnet elasticnet evtree FCNN4R FCNN4R, plyr foba gbm, plyr glmnet, Matrix h2o keras LogicReg monmlp monomvn msaenet mxnet neuralnet nnet pls qrnn RSNNS

num_param

a factor with levels alpha alpha, lambda alphas, nsteps, scale fraction hidden1, n.ensemble intercept k, lambda lambda layer1, layer2, layer3 layer1, layer2, layer3,activation1,activation2,linear.output layer1, layer2, layer3,activation1,linear.output layer1, layer2, layer3, decay layer1,layer2,layer3, decay,activation1,activation2 layer1, layer2, layer3, dropout, beta1, beta2, learningrate, activation layer1, layer2, layer3, hidden_dropout, visible_dropout layer1, layer2, layer3, learning.rate, momentum, dropout, activation ncomp nFilter,nStride,lambda, units1,units2,dropout, activation1,activation2,activation3 n.hidden, penalty, bag n.trees, interaction.depth, shrinkage, n.minobsinnode size size, decay size, decay, bag size, dropout, batch_size, lr, rho, decay, activation size, l2reg, lambda, learn_rate, momentum, gamma, minibatchsz, repeats size, lambda, batch_size, lr, rho, decay, activation sparsity treesize, ntrees units1,units2, dropout1,dropout2,batch_size, lr, rho, decay, activation1,activation2,activation3 units1,units2, l2reg, lambda, learn_rate, momentum, gamma, minibatchsz, repeats,activation1,activation2 units1,units2, l2reg, rho, activation

Details

The available model list

Source

References